+++ core/XH.js
@@ -12,7 +12,7 @@ import {elem, AppState, AppSpec, EventSupport, ReactiveSupport} from '@xh/hoist/
 import {Exception} from '@xh/hoist/exception';
 import {observable, action} from '@xh/hoist/mobx';
 import {never, wait, allSettled} from '@xh/hoist/promise';
-import {throwIf} from '@xh/hoist/utils/js';
+import {throwIf, withShortDebug} from '@xh/hoist/utils/js';
 
 import {
     ConfigService,
@@ -568,8 +568,13 @@ class XHClass {
     get acm() {return this.appContainerModel}
 
     async initServicesInternalAsync(svcs) {
-        const promises = svcs.map(it => it.initAsync()),
-            results = await allSettled(promises),
+        const promises = svcs.map(it => {
+            return withShortDebug(`Initializing ${it.constructor.name}`, () => {
+                return it.initAsync();
+            }, 'XH');
+        });
+        
+        const results = await allSettled(promises),
             errs = results.filter(it => it.state === 'rejected');
 
         if (errs.length > 0) {